requirements.txt
Flask==3.1.0
openpyxl==3.1.5
pyinstaller==6.19.0
pywebview==5.1
waitress==3.0.2
